requirementsHunyuan.txt
absl-py==2.0.0
accelerate==0.31.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.9.0
blis==0.7.10
breadability==0.1.20
Brotli==1.1.0
cachetools==5.3.1
catalogue==2.0.9
ccimport==0.4.4
certifi==2025.1.31
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.3
ConfigArgParse==1.7
contourpy==1.3.1
crypto==1.4.1
cryptography==43.0.3
cumm-cu118==0.4.11
cycler==0.12.1
cymem==2.0.8
Cython==3.1.0
dash==3.0.4
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.16.1
distro==1.9.0
Django==4.2.3
docopt==0.6.2
dtw==1.4.0
easydict==1.13
einops==0.7.0
entrypoints==0.4
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.6.0/es_core_news_md-3.6.0-py3-none-any.whl#sha256=b3d69542decdde20097a10795bddb8dc841731cd2e00e3f490306c850c2f7cd4
et-xmlfile==1.1.0
executing==2.2.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.44
glcontext==3.0.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
googlesearch-python==1.2.3
gradio==4.44.1
gradio_client==1.3.0
gradio_litmodel3d==0.0.1
greenlet==3.1.1
groovy==0.1.2
grpcio==1.59.3
gunicorn==23.0.0
h11==0.14.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.20.2
humanfriendly==10.0
humanize==4.6.0
hy3dgen==2.0.2
hyperframe==5.2.0
idna==3.10
igraph==0.11.8
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inscriptis==2.5.0
ipycanvas==0.13.3
ipydatawidgets==4.3.5
ipyevents==2.0.2
ipykernel==6.25.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==2.15.0
kiwisolver==1.4.8
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.0
langcodes==3.3.0
langsmith==0.1.143
lark==1.2.2
lazy_loader==0.4
libclang==16.0.6
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.8
ml-dtypes==0.2.0
moderngl==5.12.0
more-itertools==10.1.0
mpmath==1.3.0
msvc_runtime==14.42.34433
multidict==6.1.0
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
Naked==0.1.32
narwhals==1.37.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
omegaconf==2.3.0
onnxruntime==1.21.0
open3d==0.19.0
openai==1.54.4
openai-whisper==20230918
opencv-contrib-python==4.8.1.78
opencv-python==4.5.5.64
opencv-python-headless==4.11.0.86
openpyxl==3.1.4
opt-einsum==3.3.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathy==0.10.2
pccm==0.4.16
pdfminer.six==20231228
pdfplumber==0.11.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
plyfile==1.1
pooch==1.8.2
portalocker==3.1.1
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==6.30.2
psutil==5.9.5
psycopg2==2.9.7
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.10.5
pydantic-settings==2.6.0
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyfiglet==1.0.2
pyglet==1.5.31
pygltflib==1.16.3
Pygments==2.19.1
PyMatting==1.1.13
pymeshfix==0.17.0
pymeshlab==2023.12.post3
Pympler==1.1
PyMuPDF==1.24.12
pyparsing==3.2.3
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pythreejs==2.4.2
pytz==2025.2
pyvista==0.44.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.65
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.3.1
scipy==1.15.2
scooby==0.10.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
sentence-transformers==3.3.1
shellescape==3.8.1
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spconv-cu118==2.3.6
SpeechRecognition==3.10.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
srsly==2.4.7
stack-data==0.6.3
starlette==0.41.3
stepic==0.5.0
streamlit==1.22.0
striprtf==0.0.27
sumy==0.11.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
texttable==1.7.0
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2025.3.30
tiktoken==0.3.3
timm==1.0.15
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
transformers==4.32.1
trimesh==4.6.6
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
usd-core==25.2.post1
uvicorn==0.34.0
validators==0.34.0
vtk==9.3.1
warp-lang==1.7.0
wasabi==1.1.2
watchdog==6.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
whitenoise==6.8.2
widgetsnbextension==4.0.13
wit==6.0.1
wrapt==1.17.2
xatlas==0.0.10
yarl==1.16.0
zipp==3.21.0
